Fixed IAS' narrowing warnings
#rb pc
#rnx
#jira

[CL 27369913 by martin ridgers in ue5-main branch]

+++ Engine/Source/Runtime/Experimental/IoStoreOnDemand/Private/Governor.inl
@@ -41,7 +41,7 @@ inline uint32 FGovernorExternal::TickAllowance()
 	CycleLast = Cycle;
 
 	IPlatformFile& Ipf = IPlatformFile::GetPlatformPhysical();
-	return Ipf.GetAllowedBytesToWriteThrottledStorage();
+	return uint32(Ipf.GetAllowedBytesToWriteThrottledStorage());
 }
